Listen to global message events unless talking to a worker
Previously, we hooked `message` on the configured distant end client or server.
This is the right thing to do when talking to a worker, but doesn't work when
communicating with a frame. In that case, we want to listen for the events on
the global object. So, do the current behavior for workers and hook the global
object for everything else.

[fix #9]