Don't swallow falsy results
Previously, we checked for a response w/o a `result` or `error` by checking if
both are falsy. This is a valid (if imprecise) test for `error`. However, a
falsy `result` is valid. So, check for the presence of the property instead of
whether it's falsy.

Similarly, we decided to coerce `result` to `null` if it was falsy. Fix this in
the same manner.

Fix #15